backend/requirements.txt
fastapi[standard]~=0.115.12
pydantic~=2.11.3
python-arango~=8.1.7
python-jose[cryptography]~=3.5.0
passlib[bcrypt]~=1.7.4